1_classification_animeface/pyproject.toml
[tool.poetry]
name = "pytorch_lightning_example"
version = "0.1.0"
description = ""
authors = ["chatflip <oo.chat.flip@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.8.1"
torchvision = "^0.9.1"
hydra-core = "^1.0.6"
mlflow = "^1.15.0"
pytorch-lightning = "^1.2.6"

[tool.poetry.dev-dependencies]
pysen = "*"
black = "^20.8b1"
flake8 = "*"
isort = "*"
mypy = "*"
autopep8 = "*"
pysen-ls = "*"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
line_length = 88
py_version = "py37"

[tool.pysen.lint.source]
includes = ["."]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
